Corrected the SimpleProperty `defined` method
Removed the v0 and graphObjectAdapter sub-packages
